Modify the dotprod
The run_benchmark now prints the results in .dat files as intended by
the perf.sh sctipt.
The output has been modified to match the one expected by the scripts
especially the gnuplot one.
The Makefile has been modified to be used by the perf.sh script:
- the all target now depends on the source files (the make command now
  compile again after a touch on the main.c file)
- the clean target removes the .dat files created by the benchmarks.
The kernel may be modified with some new function that may improve
performance.