Rewrite unroll functions to handle all entries
The loops have been change to correspond to yaspr's ones.
This constitancy helps to compare the results with those obtained by
someone else (the functions being the same).
The C files would not be modify anymore (except for bug fixing) as all
the functions to be mesured have been implemented.